Initial commit of gmail-on-alexa. At the moment, it just reads all the labels in the user's gmail account.
It also oauth2 implementation to enable users to link their gmail
accounts with alexa by using a link published to alexa companion app.

User credentials are stored to dynamodb.

Since this is a prototype implementation at this point, credentials
are being logged to console for easier trouble shooting and the
dynamodb table names and schema aren't finailzed.

The redirect URL used for oauth is based on Amazon AWS API gateway.