We stop saving all tokens to dynamodb and only save the refresh_token.
The life time of an access token is pretty short (an hour usually) and so
persisting the access token actively hurts the performance because
oauth_client will first try to use the access token, fail (most of the
times except may be the very first usage of the app) and then fetch a
fresh access token.
Instead, if we simply avoid storing the access token, the client will
fetch a new access token right away and use it for the current dialog
session. This will also save us dynamodb read throughput. For that reason,
we only persist refresh_token.