So far, the authentication was completely being handled by the code. However, Alexa Skills Kit has a certain code flow where they handle account associations (the take client_id, redirect_url etc., perform necessary actions, obtain the access token, store it on their end and provide the access token to skills on every request). In this commit, we switch to that model and obtain access token from Amazon instead of storing it ourselves.
There is one big problem with this though. The refresh_token (used to
fetch a new access_token when the access_token expires) is not being
exposed by the Alexa Skills Kit. This is going to be a problem given that
access_tokens expire in as little as an hour. We need to talk to Alexa
Skills Kit team to understand how they plan to address this issue.
(exposing refresh_token isn't a big deal but their concern would probably
be around making sure that they retain control of the new access_token).